+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.6] - 2022-07-12
+
+### Added
+- `directApplicationPossible` option to job positions.
+- `benefits` to job positions.
+- Structured data for `experienceRequirements`, `responsibilities`, `jobBenefits`, `jobLocationType` and `directApplicationPossible`.
+
 ## [2.1.5] - 2022-06-30
 
 ### Fixed